+++ sdk/dotnet/Inputs/AlertChannelConfigArgs.cs
@@ -68,7 +68,7 @@ namespace Pulumi.NewRelic.Inputs
         public Input<string>? HeadersString { get; set; }
 
         /// <summary>
-        /// `0` or `1`. Flag for whether or not to attach a JSON document containing information about the associated alert to the email that is sent to recipients.
+        /// `true` or `false`. Flag for whether or not to attach a JSON document containing information about the associated alert to the email that is sent to recipients.
         /// * `webhook`
         /// </summary>
         [Input("includeJsonAttachment")]