+++ sdk/nodejs/index.ts
@@ -22,6 +22,7 @@ export * from "./getEntity";
 export * from "./getKeyTransaction";
 export * from "./infraAlertCondition";
 export * from "./nrqlAlertCondition";
+export * from "./nrqlDropRule";
 export * from "./oneDashboard";
 export * from "./provider";
 
@@ -52,6 +53,7 @@ import { EntityTags } from "./entityTags";
 import { EventsToMetricsRule } from "./eventsToMetricsRule";
 import { InfraAlertCondition } from "./infraAlertCondition";
 import { NrqlAlertCondition } from "./nrqlAlertCondition";
+import { NrqlDropRule } from "./nrqlDropRule";
 import { OneDashboard } from "./oneDashboard";
 
 const _module = {
@@ -80,6 +82,8 @@ const _module = {
                 return new InfraAlertCondition(name, <any>undefined, { urn })
             case "newrelic:index/nrqlAlertCondition:NrqlAlertCondition":
                 return new NrqlAlertCondition(name, <any>undefined, { urn })
+            case "newrelic:index/nrqlDropRule:NrqlDropRule":
+                return new NrqlDropRule(name, <any>undefined, { urn })
             case "newrelic:index/oneDashboard:OneDashboard":
                 return new OneDashboard(name, <any>undefined, { urn })
             default:
@@ -98,6 +102,7 @@ pulumi.runtime.registerResourceModule("newrelic", "index/entityTags", _module)
 pulumi.runtime.registerResourceModule("newrelic", "index/eventsToMetricsRule", _module)
 pulumi.runtime.registerResourceModule("newrelic", "index/infraAlertCondition", _module)
 pulumi.runtime.registerResourceModule("newrelic", "index/nrqlAlertCondition", _module)
+pulumi.runtime.registerResourceModule("newrelic", "index/nrqlDropRule", _module)
 pulumi.runtime.registerResourceModule("newrelic", "index/oneDashboard", _module)
 
 import { Provider } from "./provider";